1. 将项目VS2013通过VS2015自动转换 2. 去除重复定义的timespec 3. 设置lua和skynet的输出目录与工作目录一致。方便调试，否则调试时响应的配置文件或so会出现路径不对。 4. cas系列posix操作不支持除int外的操作，因此更改socket中int8->int32 5. clientsocket更新导出函数luaopen_client_socket->luaopen_clientsocket 6. client.lua更改require "clientsocket"

+++ skynet-src/socket_server.c
@@ -80,9 +80,9 @@ struct socket {
 	int64_t wb_size;
 	int fd;
 	int id;
-	uint8_t protocol;
-	uint8_t type;
-	uint16_t udpconnecting;
+	uint32_t protocol;
+	uint32_t type;
+	uint32_t udpconnecting;
 	int64_t warn_size;
 	union {
 		int size;
